QABotEngine/requirements.txt
faiss-cpu
langchain-core
langchain
langchain_openai
langchain_community
unitypredict_engines
pypdf
python-dotenv
langchain-experimental>=0.0.50